When changing values, an interruption is possible after every 5 touches.
In the case of value changes that require more than 5 screen touches, an
interruption is now possible after every 5 touches. For example, if you
have made a mistake about the DHW temperature and set it to 10 ° C, you
can correct it and the correction will take effect after 5 touches at
the latest.